Replace busy wait with interrupt-driven frame tick
Every frame tick advances the color animation one step.

The 16-bit timer that was previously used for the PWM is now
replaced by an 8-bit timer. The 16-bit timer is instead used for
generating the frame tick. This was necessary because the
required compare values for the frame tick timer would not fit
into an 8-bit register.