Refactor color blending animation
Instead of repeatedly being called in the main program, this is
now a standalone animation block that sets the desired LED
colors all by itself.

This should make it easier to add similar animations blocks
without being forced to repeating the same boilerplate code in
the main program over and over again.